Do not display volume sliders on mobile
Does not initialize volume sliders on mobile devices.  The layout object will remove any undefined objects in the right section of the control bar(in this case volume sliders) and will not represent them in control bar.  Listeners for the volume sliders are dependent on the sliders existence so they will not be assigned if the sliders do not exist.

[Finishes #93715006]